Added support for submitting Jet jobs from a reader
